+++ CHANGELOG.md
@@ -51,4 +51,4 @@
 - add new config `showUnsavedChanges.schemeTypes`
 
 ## 0.1.3
-- add new config `showUnsavedChanges.terminalPath`
+- add new config `showUnsavedChanges.terminalShellPath`